requirements.txt
boto3>=1.14
cryptography>=2.8
django==3.2.10
django-cognito-jwt
djangorestframework>=3.0
djangorestframework-camel-case>=1.2.0
django-cors-headers>=3.1
django-health-check>=3.16
django-storages>=1.9
PyJWT[crypto]<2.0.0,>=1.5.2
drf-jwt>=1.17
drf-yasg2 # This is a fork of drf-yasg which appears abandoned
packaging # Required by drf-yasg2
psycopg2==2.9.1
python-dotenv>=0.10
requests>=2.22
openpyxl==3.0
git+https://api:99WJ2_zHr4G4P3_Vxo_k@gitlab.com/scircula/size-estimator.git
git+https://api:6-ztBYRuwyUp_s4Hbf3G@gitlab.com/scircula/dataset-parser.git
tldextract
djangorestframework_camel_case
